+++ NEWS.md
@@ -1,5 +1,12 @@
 # timaR
 
+# timaR 2.9.7
+
+* Adding possibility to add internal libraries through the GUI (#159)
+* Expose more parameters to the GUI (#159)
+* Fix adducts and remove nitrogen rule
+* Updated benchmarking steps
+
 # timaR 2.9.6
 
 * Adding light-switch thanks to `pkgdown 2.1.0`.